requirements.txt
# ================================
# Core Dependencies (Required)
# ================================
# LangChain & RAG
langchain
langchain-community
langchain-chroma
langchain-ollama
chromadb
tiktoken
# Document Processing
pypdf
# Vector Database
numpy>=1.26.0,<2.0
# Configuration & Environment
python-dotenv~=1.0
PyYAML>=6.0
# Web Scraping & Search
ddgs # DuckDuckGo search (free, no API key needed)
requests~=2.26
beautifulsoup4~=4.12
lxml~=5.3
# Advanced Web Scraping (crawl4ai)
aiofiles>=24.1.0
aiohttp>=3.11.11
aiosqlite~=0.20
anyio>=4.0.0
pillow>=10.4
playwright>=1.49.0
patchright>=1.49.0
tf-playwright-stealth>=1.1.0
xxhash~=3.4
rank-bm25~=0.2
colorama~=0.4
snowballstemmer~=2.2
pydantic>=2.10
pyOpenSSL>=24.3.0
psutil>=6.1.1
nltk>=3.9.1
rich>=13.9.4
chardet>=5.2.0
brotli>=1.1.0
httpx[http2]>=0.27.2
alphashape>=1.3.1
shapely>=2.0.0
fake-useragent>=2.2.0
pdf2image>=1.17.0
PyPDF2>=3.0.1
# AI/ML
litellm>=1.53.1
# Data Processing
pandas
# Analytics (Optional)
posthog<6.0.0
# ================================
# Optional Dependencies
# ================================
# Install only if using specific features:
#
# For Google Custom Search API:
# pip install google-api-python-client
#
# google-api-python-client>=2.0.0 # Uncomment if using Google search
